requirements.txt
bcrypt-4.2.1
flask_jwt_extended-4.7.1